fix(ci): fail the hardware sync when the fetch fails, and split its tests by module
Two findings from reviewing the review.

**The weekly sync could not fail.** `roku-hardware-sync.yml` pipes `--fetch`
through `tee`, and Actions' default shell on Linux is `bash -e {0}` — no
`pipefail` — so the pipeline's exit status is `tee`'s. Every failure the
workflow's own "Failure is the point" section enumerates (an unknown RAM
spelling, a new resolution, a family carrying two RAM sizes, rokudev/dev-doc
cutting a new default branch) would exit non-zero, be masked to 0, and let the
job continue. The next step greps for `data-changed=true`, which the failure
path never prints, so it takes the else branch and reports "Roku's table is
unchanged in every field this repo derives. Nothing to open." Green, weekly,
forever — the silent staleness the generated dictionary exists to end,
reintroduced one layer up and wearing the costume of a healthy sync.

Verified rather than reasoned:

    $ bash -e -c 'false | tee /dev/null; echo "exit after pipe: $?"'
    exit after pipe: 0

`shell: bash` selects `bash --noprofile --norc -eo pipefail {0}` and fails the
step. It is the only `tee` in the repo's workflows, so nothing else is exposed.

**Three modules' tests lived in one file named for a fourth.**
`measurement-guard.test.js` had grown to 45 tests covering `measurement-guard.js`,
`roku-devices.js` and `generate/roku-hardware.js` — against the mirroring
convention `scripts/CLAUDE.md` states, in the same change set that added that
file's `scripts/data/` section. Split by subject, no test lost (29 + 10 + 6 = 45):

  - `roku-devices.test.js`  — the lookup CONTRACT: family fallback, letter-prefixed
    model numbers, null for unknown, never collapsing a two-device model number.
  - `roku-hardware.test.js` — the DATA and its parsers: invariants against the
    committed dataset, the tamper case shown failing, every cell spelling upstream uses.

1314/1314 script tests, now across 69 files.

Left open deliberately, reported and not folded in: `deviceFor()` bracket-indexes
a JSON object, so a model number colliding with an `Object.prototype` key returns
a truthy non-entry (unreachable via ECP; `Object.hasOwn` closes it), and
`home-first-paint-performance.md`'s related-files lists the dataset and the read
API but not the generator.
